implement private route

+++ src/AFS Panel/AdminPanel/pages/notification/NotificationForUser.jsx
@@ -1,14 +1,33 @@
 import { useRef, useState, useEffect } from "react";
-import { Box, Input, FormLabel, Button, Text, VStack, IconButton, HStack, Flex } from "@chakra-ui/react";
+import {
+  Box,
+  Input,
+  FormLabel,
+  Button,
+  Text,
+  VStack,
+  IconButton,
+  HStack,
+  Flex,
+} from "@chakra-ui/react";
 import JoditEditor from "jodit-react";
 import DashboardLayout from "../../components/DashboardLayout";
 import { fireDB } from "../../../firebase/FirebaseConfig";
-import { addDoc, collection, deleteDoc, doc, updateDoc } from "firebase/firestore";
+import {
+  addDoc,
+  collection,
+  deleteDoc,
+  doc,
+  updateDoc,
+} from "firebase/firestore";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchUserNotifications, selectNotifications } from "../../../redux/notifications/userNotificationsSlice";
-import DOMPurify from 'dompurify';
-import { useToast } from "@chakra-ui/react";
+import {
+  fetchUserNotifications,
+  selectNotifications,
+} from "../../../redux/notifications/userNotificationsSlice";
+import DOMPurify from "dompurify";
 import { EditIcon, DeleteIcon } from "@chakra-ui/icons";
+import { toast } from "react-toastify";
 
 const NotificationForUser = () => {
   const editor = useRef(null);
@@ -16,23 +35,16 @@ const NotificationForUser = () => {
   const [title, setTitle] = useState("");
   const [editingId, setEditingId] = useState(null);
   const dispatch = useDispatch();
-  
+
   useEffect(() => {
     dispatch(fetchUserNotifications());
   }, [dispatch]);
-  
+
   const notifications = useSelector(selectNotifications);
-  
-  const toast = useToast();
+
   const handleSave = async () => {
     if (!title || !content) {
-      toast({
-        title: "Error",
-        description: "Please enter title and content.",
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.error("Please enter title and content.");
       return;
     }
 
@@ -44,23 +56,11 @@ const NotificationForUser = () => {
       });
       setTitle("");
       setContent("");
-      toast({
-        title: "Success",
-        description: "Notification saved successfully!",
-        status: "success",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.success("Notification saved successfully!");
       dispatch(fetchUserNotifications());
     } catch (error) {
       console.error("Error saving notification: ", error);
-      toast({
-        title: "Error",
-        description: "An error occurred while saving the notification.",
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.error("An error occurred while saving the notification.");
     }
   };
 
@@ -74,37 +74,26 @@ const NotificationForUser = () => {
   };
 
   const handleDelete = async (notificationId) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete this notification?");
+    
+    if (!confirmDelete) {
+      return; // Exit the function if the user cancels the action
+    }
+    
     try {
       await deleteDoc(doc(fireDB, "userNotifications", notificationId));
-      toast({
-        title: "Success",
-        description: "Notification deleted successfully!",
-        status: "success",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.success("Notification deleted successfully!");
       dispatch(fetchUserNotifications());
     } catch (error) {
       console.error("Error deleting notification: ", error);
-      toast({
-        title: "Error",
-        description: "An error occurred while deleting the notification.",
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.error("An error occurred while deleting the notification.");
     }
   };
+  
 
   const handleUpdate = async () => {
     if (!title || !content) {
-      toast({
-        title: "Error",
-        description: "Please enter title and content.",
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.error("Please enter title and content.");
       return;
     }
 
@@ -116,39 +105,32 @@ const NotificationForUser = () => {
       setTitle("");
       setContent("");
       setEditingId(null);
-      toast({
-        title: "Success",
-        description: "Notification updated successfully!",
-        status: "success",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.success("Notification updated successfully!");
       dispatch(fetchUserNotifications());
     } catch (error) {
       console.error("Error updating notification: ", error);
-      toast({
-        title: "Error",
-        description: "An error occurred while updating the notification.",
-        status: "error",
-        duration: 5000,
-        isClosable: true,
-      });
+      toast.error("An error occurred while updating the notification.");
     }
   };
 
   return (
     <DashboardLayout title="Add User Notice">
-      <Box maxW={{md:"40vw", base:'90vw'}} mx="auto" p={6}>
+      <Box maxW={{ md: "40vw", base: "90vw" }} mx="auto" p={6}>
         <Box mb={4}>
-          <FormLabel>Title</FormLabel>
+          <FormLabel fontSize="larger" fontWeight="semibold">
+            Title
+          </FormLabel>
           <Input
+            bg="white"
             placeholder="Title"
             value={title}
             onChange={(e) => setTitle(e.target.value)}
           />
         </Box>
         <Box mb={4}>
-          <FormLabel>Description</FormLabel>
+          <FormLabel fontSize="larger" fontWeight="semibold">
+            Description
+          </FormLabel>
           <JoditEditor
             ref={editor}
             value={content}
@@ -157,7 +139,7 @@ const NotificationForUser = () => {
             onChange={(newContent) => {}}
           />
         </Box>
-        
+
         {editingId ? (
           <Button colorScheme="blue" onClick={handleUpdate}>
             Update
@@ -168,35 +150,45 @@ const NotificationForUser = () => {
           </Button>
         )}
 
-        
         <VStack mt={6} align="stretch">
-          <Text fontSize="lg" fontWeight="bold">
-            Notifications
+          <Text fontSize="x-large" fontWeight="semibold">
+            All Notifications
           </Text>
-          {notifications.map((notification) => (
-            <Box
-              key={notification.id}
-             border='1px solid gray'
-              borderRadius="md"
-              p={{base:'4', md:'2'}}
-            >
-              
-              <HStack justify='space-between'>
-                <Text>{notification.title}</Text>
-               <Flex gap={2}>
-               <IconButton
-                  icon={<EditIcon />}
-                  aria-label="Edit"
-                  onClick={() => handleEdit(notification.id)}
-                />
-                <IconButton
-                  icon={<DeleteIcon />}
-                  aria-label="Delete"
-                  onClick={() => handleDelete(notification.id)}
-                />
-               </Flex>
+          {notifications.map((notification, index) => (
+            <Box key={notification.id}>
+              <HStack
+                justify="space-between"
+                border="1px solid gray"
+                bg="white"
+                borderRadius="md"
+                p={{ base: "4", md: "2" }}
+              >
+                <Flex gap={2} align="center" fontSize="18px">
+                  <Text fontWeight="semibold">{index + 1}.</Text>
+                  <Text>{notification.title}</Text>
+                </Flex>
+                <Flex gap={2}>
+                  <IconButton
+                    icon={<EditIcon />}
+                    variant="outline"
+                    colorScheme="teal"
+                    aria-label="Edit"
+                    onClick={() => handleEdit(notification.id)}
+                  />
+                  <IconButton
+                    icon={<DeleteIcon />}
+                    colorScheme="red"
+                    aria-label="Delete"
+                    onClick={() => handleDelete(notification.id)}
+                  />
+                </Flex>
               </HStack>
-              <Text mt={5} dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(notification.content) }} />
+              {/* <Text
+                mt={5}
+                dangerouslySetInnerHTML={{
+                  __html: DOMPurify.sanitize(notification.content),
+                }}
+              /> */}
             </Box>
           ))}
         </VStack>
